pyproject.toml
[project]
name = "outlyne"
version = "0.1.0"
description = "AI-powered sketch-to-image meta search engine with sub-second perceived latency."
readme = "README.md"
authors = [
    { name = "Rohit", email = "dv6510@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "blake3>=1.0.8",
    "bs4>=0.0.2",
    "cachetools>=6.2.4",
    "ddgs>=9.10.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "nncf>=2.19.0",
    "numpy>=1.26.4",
    "onnxruntime>=1.23.2",
    "opencv-python-headless>=4.11.0.86",
    "openvino>=2025.0.0",
    "openvino-telemetry>=2025.2.0",
    "optimum>=2.1.0",
    "optimum-intel[openvino]>=1.22.0",
    "pillow>=12.1.0",
    "protobuf>=6.33.4",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sentencepiece>=0.2.1",
    "torch>=2.9.0",
    "torchvision>=0.21.0",
    "transformers>=4.48.3",
    "uvicorn>=0.40.0",
]

[project.scripts]
outlyne = "main:main"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "PT"]
ignore = []

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    '\.venv',
    '\.cache',
    'node_modules',
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
